fix: move quarterFilter to its own filter slice instead of search slice
